[openstack-seeder] Seed custom traits from flavors
We have started to use custom traits for flavors and hosts more. This
led to the situation that after the flavors where seeded the trait was
not yet existing in the region. This fails _any_ placement of a VM
with a flavor that mentions a trait in the extra-specs, even if it
only sets the trait as "forbidden".

Openstack-Placement cannot act on any trait spec if it does not know
the trait and will fail all deployments with traited flavors, required
or forbidden.

To fix this, collect any and all custom traits specified in the flavor
seed and create them. This works exactly like the seeding of custom
resource classes mentioned in the flavors' extra-specs.

Note: Unlike the placement API request for a resource-class which
needs API version 1.7 or higher (previous versions failed on an
existing resource-class), the trait API endpoint never had this
limitation. Thus the trait-seed API call does not pin the API version.