Don't seed flavors whose required traits aren't associated
If there isn't any resource provider that has a required trait for a
flavor, servers with that flavor will always fail to schedule.

This has been happening with the new HANA flavors which are the first
to be hardware-dependent, and thus not homogeneously available in all
regions.

A trait in Openstack/Placement might be absent (not yet created and
thus not known to Openstack at all), created but yet unassociated
with a resource provider, or associated. When a trait is absent, a
deployment with that flavor will always fail, even if the trait
is "forbidden" by the flavor (i.e. must not be set by a resource
provider to be eligible).

The seeder now checks both which traits are present and which traits
are associated. It then seeds absent traits, and only seeds those
flavors who have no required traits or whose required traits have at
least one resource provider with it. Operators then have time to
associate new traits with resource providers, and when the seeder
runs the next time, any flavors that require those traits to be
associated will then automatically be seeded.

In effect, the availability of HANA-flavor-capable clusters will now
be readily apparent by running `openstack flavor list`.