Fix flavor extra_specs seeding
In previous commits we added seeding flavors only if their traits are
assigned. We got those traits from the extra_specs. In this process, we
removed the extra_specs key from the flavor and thus, when the flavor
was deemed ready for seeding, we did not have any extra_specs for
seeding anymore.